correct a bug with the postgres notifier
Once the job runner worker fork has forked all the database connections
are now all subject to fork-safety code and will be cleaned up.
Unfortunately the Mojo::Pg::PubSub::_cleanup code isn't run until a
database connection is needed. However, since no db access is needed
from the job side the cleanup code is not run thereby leaving a
connection listening for notifications for the wrong process. Once a
notification happens, say when the task manually calls finish, the bad
connection is used and things go badly from there.

This commit adds a quick and dirty cleanup for the pubsub object. By
unsubscribing from a dummy channel the db connection is inspected and is
thus noticed to be from the old process and cleaned up.

If Mojo::Pg::PubSub should in the future provide a public method to
cleanup the connection, it should be used. If Minion::Job and
Minion::Backend::Pg should in the future ensure that the
Mojo::Pg::PubSub instance is cleaned up then all of this handling can be
removed.